Move create entities methods to common TestData
